splitting the fdrs_functions
Also removed most of the FDRS_* files
